+++ react-ui/src/components/SearchBar/SearchBar.scss
@@ -1,4 +1,5 @@
 .search-component {
+    text-align: center;
     background: #222 url('./restaurant-alcohol-bar-drinks.jpg') center center no-repeat;
     min-height: 50rem;
     display: flex;
