+++ react-ui/src/components/SearchBar/index.js
@@ -1,4 +1,8 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
+import { withRouter } from 'react-router-dom'
+import { selectLocation, fetchBarsIfNeeded, invalidateLocation } from '../../actions'
 import { Button, Form, FormGroup, FormControl } from 'react-bootstrap';
 
 import './SearchBar.css';
@@ -19,9 +23,19 @@ class SearchBar extends Component {
 
     handleSubmit(e) {
         e.preventDefault()
+        this.props.history.push(
+            {
+                pathname: '/bars'
+            }
+        )
         if (this.state.searchValue.length !== 0) {
-            this.props.submitSearch(this.state.searchValue)
-            this.setState({searchValue: ''})
+            const { dispatch, selectedLocation } = this.props
+            dispatch(selectLocation(this.state.searchValue))
+            dispatch(invalidateLocation(selectedLocation))
+            if (selectedLocation.length !== 0) {
+                dispatch(fetchBarsIfNeeded(selectedLocation))
+                this.setState({searchValue: ''})
+            }
         }
     }
 
@@ -47,4 +61,27 @@ class SearchBar extends Component {
     }
 }
 
-export default SearchBar;
+SearchBar.propTypes = {
+    selectedLocation: PropTypes.string.isRequired,
+    bars: PropTypes.array.isRequired,
+    lastUpdated: PropTypes.number,
+    dispatch: PropTypes.func.isRequired
+}
+
+const mapStateToProps = state => {
+    const { selectedLocation, barsByLocation } = state
+    const {
+        lastUpdated,
+        items: bars
+    } = barsByLocation[selectedLocation] || {
+        items: []
+    }
+
+    return {
+        selectedLocation,
+        bars,
+        lastUpdated
+    }
+}
+
+export default withRouter(connect(mapStateToProps)(SearchBar));
