+++ react-ui/src/modules/App/index.js
@@ -1,61 +1,18 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
-import { selectLocation, fetchBarsIfNeeded, invalidateLocation } from '../../actions'
-import Navigation from './components/Navigation';
-import SearchBar from './components/SearchBar';
+import SearchBar from '../../components/SearchBar';
 
 import './App.css';
 
 class App extends Component {
-    constructor(props) {
-        super(props)
-        this.submitSearch = this.submitSearch.bind(this)
-    }
-
-    submitSearch(searchLocation) {
-        const { dispatch, selectedLocation } = this.props
-        dispatch(selectLocation(searchLocation))
-        dispatch(invalidateLocation(selectedLocation))
-        if (selectedLocation.length !== 0) {
-            dispatch(fetchBarsIfNeeded(selectedLocation))
-        }
-    }
-
     render() {
         return (
             <div className='App'>
-                <Navigation />
-                <SearchBar submitSearch={this.submitSearch} />
+                <SearchBar />
                 <div className='App-header'>
                 </div>
-                {this.props.children}
             </div>
         )
     }
 }
 
-App.propTypes = {
-    selectedLocation: PropTypes.string.isRequired,
-    bars: PropTypes.array.isRequired,
-    lastUpdated: PropTypes.number,
-    dispatch: PropTypes.func.isRequired
-}
-
-const mapStateToProps = state => {
-    const { selectedLocation, barsByLocation } = state
-    const {
-        lastUpdated,
-        items: bars
-    } = barsByLocation[selectedLocation] || {
-        items: []
-    }
-
-    return {
-        selectedLocation,
-        bars,
-        lastUpdated
-    }
-}
-
-export default connect(mapStateToProps)(App);
+export default App;
