+++ react-ui/src/modules/Bars/Bars.scss
@@ -1,3 +1,7 @@
+.bars-container {
+    text-align: center;
+}
+
 .container {
     max-width: 1100px;
     margin: 0 auto;
