--- react-ui/src/modules/Home/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const Home = (props) => {
-    return (
-        <div>
-            <h1>Home!</h1>
-            <p>Welcome to Home!</p>
-            {props.children}
-        </div>
-    )
-}
-
-export default Home;
