+++ react-ui/src/routes.js
@@ -1,20 +1,21 @@
 import React, { PropTypes } from 'react';
 import { Provider } from 'react-redux';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 
+import Navigation from './modules/Navigation';
 import App from './modules/App';
-import Home from './modules/Home';
+import Login from './modules/Login';
 import Bars from './modules/Bars';
 
 const Routes = ({ store }) => (
     <Provider store={store}>
         <Router>
-            <App>
-                <Switch>
-                    <Route exact path='/' component={Home} />
-                    <Route path='/bars' component={Bars} />
-                </Switch>
-            </App>
+            <div className='router'>
+                <Navigation />
+                <Route exact path='/' component={App} />
+                <Route path='/login' component={Login} />
+                <Route path='/bars' component={Bars} />
+            </div>
         </Router>
     </Provider>
 )
